+++ integration-test/src/test/java/org/apache/iotdb/confignode/it/partition/IoTDBPartitionGetterIT.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.confignode.it;
+package org.apache.iotdb.confignode.it.partition;
 
 import org.apache.iotdb.common.rpc.thrift.TConsensusGroupId;
 import org.apache.iotdb.common.rpc.thrift.TConsensusGroupType;
@@ -75,19 +75,20 @@ import static org.apache.iotdb.confignode.it.utils.ConfigNodeTestUtils.generateP
 
 @RunWith(IoTDBTestRunner.class)
 @Category({ClusterIT.class})
-public class IoTDBClusterPartitionIT {
+public class IoTDBPartitionGetterIT {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(IoTDBClusterPartitionIT.class);
+  private static final Logger LOGGER = LoggerFactory.getLogger(IoTDBPartitionGetterIT.class);
 
-  protected static String originalConfigNodeConsensusProtocolClass;
-  protected static String originalSchemaRegionConsensusProtocolClass;
-  protected static String originalDataRegionConsensusProtocolClass;
+  private static String originalConfigNodeConsensusProtocolClass;
+  private static String originalSchemaRegionConsensusProtocolClass;
+  private static String originalDataRegionConsensusProtocolClass;
+  private static final String testConsensusProtocolClass = ConsensusFactory.RATIS_CONSENSUS;
 
-  protected static int originalSchemaReplicationFactor;
-  protected static int originalDataReplicationFactor;
+  private static int originalSchemaReplicationFactor;
+  private static int originalDataReplicationFactor;
   private static final int testReplicationFactor = 3;
 
-  protected static long originalTimePartitionInterval;
+  private static long originalTimePartitionInterval;
   private static final long testTimePartitionInterval = 604800000;
 
   private static final String sg = "root.sg";
@@ -105,10 +106,10 @@ public class IoTDBClusterPartitionIT {
         ConfigFactory.getConfig().getSchemaRegionConsensusProtocolClass();
     originalDataRegionConsensusProtocolClass =
         ConfigFactory.getConfig().getDataRegionConsensusProtocolClass();
-    ConfigFactory.getConfig().setConfigNodeConsesusProtocolClass(ConsensusFactory.RATIS_CONSENSUS);
+    ConfigFactory.getConfig().setConfigNodeConsesusProtocolClass(testConsensusProtocolClass);
     ConfigFactory.getConfig()
-        .setSchemaRegionConsensusProtocolClass(ConsensusFactory.RATIS_CONSENSUS);
-    ConfigFactory.getConfig().setDataRegionConsensusProtocolClass(ConsensusFactory.RATIS_CONSENSUS);
+        .setSchemaRegionConsensusProtocolClass(testConsensusProtocolClass);
+    ConfigFactory.getConfig().setDataRegionConsensusProtocolClass(testConsensusProtocolClass);
 
     originalSchemaReplicationFactor = ConfigFactory.getConfig().getSchemaReplicationFactor();
     originalDataReplicationFactor = ConfigFactory.getConfig().getDataReplicationFactor();
@@ -118,7 +119,8 @@ public class IoTDBClusterPartitionIT {
     originalTimePartitionInterval = ConfigFactory.getConfig().getTimePartitionInterval();
     ConfigFactory.getConfig().setTimePartitionIntervalForRouting(testTimePartitionInterval);
 
-    EnvFactory.getEnv().initBeforeClass();
+    // Init 1C3D environment
+    EnvFactory.getEnv().initClusterEnvironment(1, 3);
     prepareData();
   }
 
@@ -348,201 +350,6 @@ public class IoTDBClusterPartitionIT {
           }
         }
       }
-    }
-  }
-
-  // TODO: Optimize this in IOTDB-4334
-  @Test
-  public void testPartitionDurable() throws IOException, TException, InterruptedException {
-    final int testDataNodeId = 0;
-    final int seriesPartitionBatchSize = 10;
-    final int timePartitionBatchSize = 10;
-
-    // Shutdown the first DataNode
-    EnvFactory.getEnv().shutdownDataNode(testDataNodeId);
-
-    try (SyncConfigNodeIServiceClient client =
-        (SyncConfigNodeIServiceClient) EnvFactory.getEnv().getLeaderConfigNodeConnection()) {
-      final String sg0 = sg + 0;
-      final String sg1 = sg + 1;
-
-      // Test getOrCreateDataPartition, ConfigNode should create DataPartition and return
-      Map<String, Map<TSeriesPartitionSlot, List<TTimePartitionSlot>>> partitionSlotsMap =
-          ConfigNodeTestUtils.constructPartitionSlotsMap(
-              sg0,
-              0,
-              seriesPartitionBatchSize,
-              0,
-              timePartitionBatchSize,
-              testTimePartitionInterval);
-      TDataPartitionReq dataPartitionReq = new TDataPartitionReq(partitionSlotsMap);
-      TDataPartitionTableResp dataPartitionTableResp = null;
-      for (int retry = 0; retry < 5; retry++) {
-        // Build new Client since it's unstable in Win8 environment
-        try (SyncConfigNodeIServiceClient configNodeClient =
-            (SyncConfigNodeIServiceClient) EnvFactory.getEnv().getLeaderConfigNodeConnection()) {
-          dataPartitionTableResp = configNodeClient.getOrCreateDataPartitionTable(dataPartitionReq);
-          if (dataPartitionTableResp != null) {
-            break;
-          }
-        } catch (Exception e) {
-          // Retry sometimes in order to avoid request timeout
-          LOGGER.error(e.getMessage());
-          TimeUnit.SECONDS.sleep(1);
-        }
-      }
-      Assert.assertNotNull(dataPartitionTableResp);
-      Assert.assertEquals(
-          TSStatusCode.SUCCESS_STATUS.getStatusCode(),
-          dataPartitionTableResp.getStatus().getCode());
-      Assert.assertNotNull(dataPartitionTableResp.getDataPartitionTable());
-      ConfigNodeTestUtils.checkDataPartitionTable(
-          sg0,
-          0,
-          seriesPartitionBatchSize,
-          0,
-          timePartitionBatchSize,
-          testTimePartitionInterval,
-          dataPartitionTableResp.getDataPartitionTable());
-
-      // Check Region count
-      int runningCnt = 0;
-      int unknownCnt = 0;
-      TShowRegionResp showRegionResp = client.showRegion(new TShowRegionReq());
-      Assert.assertEquals(
-          TSStatusCode.SUCCESS_STATUS.getStatusCode(), showRegionResp.getStatus().getCode());
-      for (TRegionInfo regionInfo : showRegionResp.getRegionInfoList()) {
-        if (RegionStatus.Running.getStatus().equals(regionInfo.getStatus())) {
-          runningCnt += 1;
-        } else if (RegionStatus.Unknown.getStatus().equals(regionInfo.getStatus())) {
-          unknownCnt += 1;
-        }
-      }
-      // The runningCnt should be exactly twice as the unknownCnt
-      // since there exists one DataNode is shutdown
-      Assert.assertEquals(unknownCnt * 2, runningCnt);
-
-      // Wait for shutdown check
-      TShowClusterResp showClusterResp;
-      while (true) {
-        boolean containUnknown = false;
-        showClusterResp = client.showCluster();
-        for (TDataNodeLocation dataNodeLocation : showClusterResp.getDataNodeList()) {
-          if (NodeStatus.Unknown.getStatus()
-              .equals(showClusterResp.getNodeStatus().get(dataNodeLocation.getDataNodeId()))) {
-            containUnknown = true;
-            break;
-          }
-        }
-        if (containUnknown) {
-          break;
-        }
-      }
-      runningCnt = 0;
-      unknownCnt = 0;
-      showClusterResp = client.showCluster();
-      for (TDataNodeLocation dataNodeLocation : showClusterResp.getDataNodeList()) {
-        if (NodeStatus.Running.getStatus()
-            .equals(showClusterResp.getNodeStatus().get(dataNodeLocation.getDataNodeId()))) {
-          runningCnt += 1;
-        } else if (NodeStatus.Unknown.getStatus()
-            .equals(showClusterResp.getNodeStatus().get(dataNodeLocation.getDataNodeId()))) {
-          unknownCnt += 1;
-        }
-      }
-      Assert.assertEquals(2, runningCnt);
-      Assert.assertEquals(1, unknownCnt);
-
-      // Test getOrCreateDataPartition, ConfigNode should create DataPartition and return
-      partitionSlotsMap =
-          ConfigNodeTestUtils.constructPartitionSlotsMap(
-              sg1,
-              0,
-              seriesPartitionBatchSize,
-              0,
-              timePartitionBatchSize,
-              testTimePartitionInterval);
-      dataPartitionReq = new TDataPartitionReq(partitionSlotsMap);
-      for (int retry = 0; retry < 5; retry++) {
-        // Build new Client since it's unstable in Win8 environment
-        try (SyncConfigNodeIServiceClient configNodeClient =
-            (SyncConfigNodeIServiceClient) EnvFactory.getEnv().getLeaderConfigNodeConnection()) {
-          dataPartitionTableResp = configNodeClient.getOrCreateDataPartitionTable(dataPartitionReq);
-          if (dataPartitionTableResp != null) {
-            break;
-          }
-        } catch (Exception e) {
-          // Retry sometimes in order to avoid request timeout
-          LOGGER.error(e.getMessage());
-          TimeUnit.SECONDS.sleep(1);
-        }
-      }
-      Assert.assertNotNull(dataPartitionTableResp);
-      Assert.assertEquals(
-          TSStatusCode.SUCCESS_STATUS.getStatusCode(),
-          dataPartitionTableResp.getStatus().getCode());
-      Assert.assertNotNull(dataPartitionTableResp.getDataPartitionTable());
-      ConfigNodeTestUtils.checkDataPartitionTable(
-          sg1,
-          0,
-          seriesPartitionBatchSize,
-          0,
-          timePartitionBatchSize,
-          testTimePartitionInterval,
-          dataPartitionTableResp.getDataPartitionTable());
-
-      // Check Region count and status
-      runningCnt = 0;
-      unknownCnt = 0;
-      showRegionResp = client.showRegion(new TShowRegionReq());
-      Assert.assertEquals(
-          TSStatusCode.SUCCESS_STATUS.getStatusCode(), showRegionResp.getStatus().getCode());
-      for (TRegionInfo regionInfo : showRegionResp.getRegionInfoList()) {
-        if (RegionStatus.Running.getStatus().equals(regionInfo.getStatus())) {
-          runningCnt += 1;
-        } else if (RegionStatus.Unknown.getStatus().equals(regionInfo.getStatus())) {
-          unknownCnt += 1;
-        }
-      }
-      // The runningCnt should be exactly twice as the unknownCnt
-      // since there exists one DataNode is shutdown
-      Assert.assertEquals(unknownCnt * 2, runningCnt);
-
-      EnvFactory.getEnv().startDataNode(testDataNodeId);
-      // Wait for heartbeat check
-      while (true) {
-        boolean containUnknown = false;
-        showClusterResp = client.showCluster();
-        for (TDataNodeLocation dataNodeLocation : showClusterResp.getDataNodeList()) {
-          if (NodeStatus.Unknown.getStatus()
-              .equals(showClusterResp.getNodeStatus().get(dataNodeLocation.getDataNodeId()))) {
-            containUnknown = true;
-            break;
-          }
-        }
-        if (!containUnknown) {
-          break;
-        }
-      }
-
-      // All Regions should alive after the testDataNode is restarted
-      boolean allRunning = true;
-      for (int retry = 0; retry < 30; retry++) {
-        allRunning = true;
-        showRegionResp = client.showRegion(new TShowRegionReq());
-        for (TRegionInfo regionInfo : showRegionResp.getRegionInfoList()) {
-          if (!RegionStatus.Running.getStatus().equals(regionInfo.getStatus())) {
-            allRunning = false;
-            break;
-          }
-        }
-        if (allRunning) {
-          break;
-        }
-
-        TimeUnit.SECONDS.sleep(1);
-      }
-      Assert.assertTrue(allRunning);
     }
   }
